+++ HopperRender/resource.h
@@ -29,7 +29,7 @@
 #define IDC_SIDEBYSIDE1                 1030
 #define IDC_SIDEBYSIDE2                 1031
 #define IDC_DEFAULTS                    1032
-#define IDC_USEDISPLAYFPS               1033
+#define IDC_FRAMERATEMODE               1033
 #define IDC_AVGOFCCALCTIME              1034
 #define IDC_PEAKOFCCALCTIME             1035
 #define IDC_TOTALFRAMEDELTA             1036